Feature/contacts page

+++ public/js/deleteContact.js
@@ -7,16 +7,18 @@ const deleteSelected = async (event) => {
   console.log("click")
   const checkedCheckboxes = document.querySelectorAll('input[type="checkbox"]:checked');
   for (let i = 0; i < checkedCheckboxes.length; i++) {
-    const idNumber = checkedCheckboxes[i].dataset.number
-    const response = await fetch(`/api/contact/${idNumber}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
-    checkedCheckboxes[i].checked=false;
-    if (!response.ok) {
-      alert('Failed to create contact');
+    if (checkedCheckboxes[i].id !== "checkboxall") {
+      const idNumber = checkedCheckboxes[i].dataset.number
+      const response = await fetch(`/api/contact/${idNumber}`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      });
+      checkedCheckboxes[i].checked = false;
+      if (!response.ok) {
+        alert('Failed to create contact');
+      }
     }
   }
   window.location.reload();
